requirements.txt
Flask==3.0.0
requests==2.31.0
Werkzeug==3.0.1
python-dotenv==1.0.0
SQLAlchemy==2.0.23
chromadb==0.4.22
sentence-transformers==2.2.2
python-dateutil==2.8.2
email-validator==2.1.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5